Migrate the project to typescript
- added prettier
- tailwind/ui plugin
