+++ .vscode/settings.json
@@ -1,0 +1,11 @@
+{
+  "editor.formatOnSave": true,
+  "editor.detectIndentation": false,
+  "editor.tabSize": 2,
+  "cSpell.words": [
+    "Visualbi",
+    "mediumtext",
+    "shortcode",
+    "varchar"
+  ]
+}
