+++ Webpack/webpack.prod.js
@@ -3,8 +3,8 @@ const common = require('./webpack.common.js');
 const HtmlWebPackPlugin = require("html-webpack-plugin");
 
 module.exports = merge(common, {
-    "mode": "production",
-    output: {
-      filename: '[name].[contentHash].bundle.js'
-    }
+  "mode": "production",
+  output: {
+    filename: '[name].[contentHash].bundle.js'
+  }
 });
