--- src/Components/Logo.js
@@ -1,9 +1,0 @@
-import React from "react";
-
-const Logo = (props) => {
-    return (<div className="flex-shrink-0 flex items-center px-4">
-        <img className="h-8 w-auto" src="/img/logos/workflow-logo-on-brand.svg" alt="Workflow" />
-    </div>);
-};
-
-export default Logo;
