+++ src/Styles/App.css
@@ -3,4 +3,3 @@
 @import "tailwindcss/components";
 
 @import "tailwindcss/utilities";
-
