+++ tailwind.config.js
@@ -1,21 +1,5 @@
 module.exports = {
-  important: true,
-  theme: {
-    fontFamily: {
-      display: ['Gilroy', 'sans-serif'],
-      body: ['Graphik', 'sans-serif'],
-    },
-    extend: {
-      colors: {
-        cyan: '#9cdbff',
-      },
-      margin: {
-        '96': '24rem',
-        '128': '32rem',
-      },
-    }
-  },
-  variants: {
-    opacity: ['responsive', 'hover']
-  }
+  plugins: [
+    require('@tailwindcss/ui'),
+  ]
 };
